# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Empleo",
        "description": "Sistema de gestión de ofertas de empleo y seguimiento de candidatos.",
        "contact": {
            "email": "admin@centro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/api/bajas/ejecutar": {
            "post": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Darse de baja del portal (Usuario)",
                "description": "El usuario se desactiva a sí mismo. Valida que no tenga procesos activos.",
                "operationId": "5efe56b3ce2350ef5be17ae32f973817",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo_baja_id"
                                ],
                                "properties": {
                                    "motivo_baja_id": {
                                        "type": "integer"
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cuenta desactivada y sesión cerrada"
                    },
                    "422": {
                        "description": "Error: Tiene ofertas o candidaturas activas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bajas/historial": {
            "get": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Historial de bajas (Admin)",
                "operationId": "1c88b3e7d383ae5dd2440f14f3eea3c5",
                "parameters": [
                    {
                        "name": "busqueda",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de usuarios inactivos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bajas/motivos": {
            "post": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Crear nuevo motivo de baja (Admin)",
                "operationId": "dd06f2f1c5bff6cd6005afe145551029",
                "responses": {
                    "201": {
                        "description": "Motivo creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bajas/motivos/{id}": {
            "delete": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Eliminar o desactivar un motivo de baja",
                "description": "Si el motivo tiene historial, solo se desactiva. Si no, se podría eliminar.",
                "operationId": "20ec22f76a9e3b14fad07cb77b6b103c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acción realizada correctamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Actualizar un motivo de baja",
                "operationId": "f8b4d7bec3ec9aa5854c8e121ad6c0bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "motivo": {
                                        "type": "string",
                                        "example": "Motivo actualizado"
                                    },
                                    "visible_alumno": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "visible_empresa": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motivo actualizado correctamente"
                    },
                    "404": {
                        "description": "Motivo no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bajas/forzosa/{idUsuario}": {
            "post": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Baja administrativa de un usuario",
                "description": "Desactiva al usuario, cierra sus ofertas (si es empresa) o retira candidaturas (si es alumno).",
                "operationId": "8687974013fa0e62b2534a0f8c46a792",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo_baja_id"
                                ],
                                "properties": {
                                    "motivo_baja_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "comentario_baja": {
                                        "type": "string",
                                        "example": "Baja por incumplimiento de normas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario y procesos gestionados correctamente"
                    },
                    "403": {
                        "description": "No puedes darte de baja a ti mismo"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/usuarios/{idUsuario}/reactivar": {
            "patch": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Reactivar un usuario inactivo (Admin)",
                "operationId": "dbb22ac551a7015744bb9b2f030a91f4",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario reactivado y datos de baja limpiados"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/usuarios/{idUsuario}/reset-password": {
            "patch": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Resetear contraseña de usuario (Admin)",
                "description": "Genera una clave de 8 caracteres y activa el flag change_pass.",
                "operationId": "be8b08f3313e5078be07652f98c1dd09",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clave temporal generada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/usuarios/actualizar-password": {
            "patch": {
                "tags": [
                    "Auth"
                ],
                "summary": "Actualizar contraseña por cambio forzoso (Usuario)",
                "description": "Limpia el flag change_pass tras la actualización exitosa.",
                "operationId": "a3ed360765867b193d95d3d0b2a8f751",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "Authorization",
                "in": "header",
                "example": "Bearer 19|tBfCzyfhD1Ib3VA5HPyHfR6CjrIoyT5VPq7s0mMnb19cdfbe"
            }
        }
    },
...
}